Shorten the test
git-svn-id: https://svn.apache.org/repos/asf/activemq/trunk@1490034 13f79535-47bb-0310-9956-ffa450edef68

+++ activemq-client/src/test/java/org/apache/activemq/ActiveMQMessageAuditNotSyncTest.java
@@ -48,7 +48,7 @@ public class ActiveMQMessageAuditNotSyncTest {
     @Test
     public void testAuditDepth() {
 
-        int maxAuditDepth = Integer.MAX_VALUE;
+        int maxAuditDepth = Short.MAX_VALUE;
 
         ConnectionId connectionId = new ConnectionId(connectionIdGenerator.generateId());
         SessionId sessionId = new SessionId(connectionId, sessionIdGenerator.getNextSequenceId());
